backend/pyproject.toml
[project]
name = "drbfm_assist"
version = "0.0.1"
description = "Starter template for making a new agent LangGraph."
authors = [
    { name = "Amane Suzuki", email = "amane_suzuki@caddi.com" },
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "pandas>=2.0.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.65.0",
    "asyncio>=3.4.3",
    "openai>=1.97.1",
    "click>=8.2.1",
    "pydantic>=2.11.7",
    "langgraph>=0.6.2",
    "langfuse>=3.2.1",
    "langchain>=0.3.27",
    "google-genai>=0.7.0",
    "google-cloud-aiplatform>=1.60.0",
    "google-auth>=2.0.0",
    "openinference-instrumentation-google-genai>=0.1.5",
    "tenacity>=8.2.0",
    "pyppeteer>=0.0.25",
    "scikit-learn>=1.3.0",
    "db-dtypes>=1.4.3",
    "langsmith>=0.4.10",
    "pyjwt>=2.8.0",
    "langgraph-sdk>=0.2.9",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["drassist"]
[tool.setuptools.package-dir]
"drassist" = "drassist"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "ipython>=9.5.0",
    "langchain-google-vertexai>=2.0.28",
    "langgraph-cli[inmem]>=0.2.8",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]
